pyproject.toml
[project]
name = "spriter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "notebook>=7.4.4",
    "pillow>=11.3.0",
    "pyxelate",
]

[tool.uv.sources]
pyxelate = { git = "https://github.com/sedthh/pyxelate.git" }
